Port logging to library-init
